Chore: MemberDAO에 있던 method BoardDAO로 이동되었습니다.
정상작동됩니다.
