+++ semi-project-moco/src/main/java/org/kosta/moco/model/BoardDAO.java
@@ -212,4 +212,40 @@ public class BoardDAO {
 		
 		return pvo;
 	}
+	
+	
+	public  ArrayList<PostVO> getMemberPosts(String email) throws SQLException {
+		ArrayList<PostVO> list=new ArrayList<PostVO>();
+		Connection con=null;
+		PreparedStatement pstmt=null;
+		ResultSet rs=null;
+		try {
+			con=dataSource.getConnection();
+			StringBuilder sql = new StringBuilder();
+			sql.append("select v.post_no, v.post_title, v.post_regdate, v.hits, l.language ");
+			sql.append("from ( ");
+			sql.append("select b.post_no, b.post_title, b.post_regdate, b.hits, b.language_code as lang ");
+			sql.append("from moco_qna_board b, moco_member m where m.email =? ");
+			sql.append(") v, moco_service_language l ");
+			sql.append("where v.lang = l.language_code ");
+			sql.append("order by v.post_no desc ");
+			pstmt=con.prepareStatement(sql.toString());
+			pstmt.setString(1, email);
+			rs=pstmt.executeQuery();
+			while(rs.next()) {
+				PostVO pvo=new PostVO();
+				pvo.setPost_no(rs.getInt("post_no"));
+				pvo.setPost_title(rs.getString("post_title"));
+				pvo.setPost_regdate(rs.getString("post_regdate"));
+				pvo.setHits(rs.getInt("hits"));
+				LanguageVO lvo = new LanguageVO();
+				lvo.setLanguage(rs.getString("language"));
+				pvo.setLvo(lvo);
+				list.add(pvo);
+			}
+		}finally {
+			closeAll(rs, pstmt, con);
+		}
+		return list;
+	}
 }
